scripts/requirements.txt
GitPython==3.1.44
Requests==2.32.4
